Move running related code to separate crate
