+++ Cargo.toml
@@ -2,6 +2,7 @@
 resolver = "2"
 members = [
     "crates/forge",
+    "crates/forge_runner",
     "crates/test-collector",
     "crates/cast",
     "crates/cheatnet",
