+++ crates/forge/benches/forge_bench/compile_tests.rs
@@ -1,7 +1,8 @@
 use crate::forge_bench::collect_tests;
 use assert_fs::TempDir;
 use camino::Utf8PathBuf;
-use forge::{CrateLocation, TestCompilationTarget};
+use forge::collecting::TestCompilationTarget;
+use forge::CrateLocation;
 use std::path::PathBuf;
 use std::str::FromStr;
 use test_collector::LinkedLibrary;
