+++ crates/forge/src/collecting.rs
@@ -1,17 +1,13 @@
-use crate::{CrateLocation, RunnerParams, BUILTINS};
+use crate::{CrateLocation, RunnerParams};
 use anyhow::{anyhow, Context, Result};
 use cairo_lang_sierra::program::Program;
 use camino::{Utf8Path, Utf8PathBuf};
+use forge_runner::BUILTINS;
 use rayon::prelude::{IntoParallelRefIterator, ParallelIterator};
-use starknet::core::types::BlockId;
 use test_collector::{collect_tests, ForkConfig, LinkedLibrary, RawForkConfig, TestCase};
-use url::Url;
 use walkdir::WalkDir;
 
 pub(crate) type CompiledTestCrateRaw = CompiledTestCrate<RawForkConfig>;
-pub(crate) type CompiledTestCrateRunnable = CompiledTestCrate<ValidatedForkConfig>;
-
-pub(crate) type TestCaseRunnable = TestCase<ValidatedForkConfig>;
 
 #[derive(Debug, Clone)]
 pub struct CompiledTestCrate<T: ForkConfig> {
@@ -19,14 +15,6 @@ pub struct CompiledTestCrate<T: ForkConfig> {
     pub test_cases: Vec<TestCase<T>>,
     pub tests_location: CrateLocation,
 }
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ValidatedForkConfig {
-    pub url: Url,
-    pub block_id: BlockId,
-}
-
-impl ForkConfig for ValidatedForkConfig {}
 
 #[derive(Debug, PartialEq)]
 pub struct TestCompilationTarget {
