+++ crates/forge/src/lib.rs
@@ -1,167 +1,21 @@
-use std::collections::HashMap;
-use std::fmt::Debug;
-
-use anyhow::{anyhow, Context, Result};
-use ark_std::iterable::Iterable;
-use cairo_felt::Felt252;
-use camino::{Utf8Path, Utf8PathBuf};
-
-use futures::StreamExt;
-use running::blocking_run_from_test;
-use tokio::sync::mpsc::{channel, Sender};
-
-use std::sync::Arc;
-use test_case_summary::TestCaseSummary;
-use tokio::task::{self, JoinHandle};
-use tokio_util::sync::CancellationToken;
-
-use cairo_lang_runner::SierraCasmRunner;
-use cairo_lang_sierra::ids::ConcreteTypeId;
-use cairo_lang_sierra::program::Function;
-use cairo_lang_sierra_to_casm::metadata::MetadataComputationConfig;
-use cairo_lang_utils::ordered_hash_map::OrderedHashMap;
-use futures::stream::FuturesUnordered;
+use anyhow::{anyhow, Result};
+use camino::Utf8Path;
 use itertools::Itertools;
+use std::fmt::Debug;
+use std::sync::Arc;
 
-use once_cell::sync::Lazy;
-use smol_str::SmolStr;
+use forge_runner::test_crate_summary::TestCrateSummary;
+use forge_runner::TestCase as RunnerTestCase;
+use forge_runner::{CancellationTokens, ForkConfig, RunnerConfig, RunnerParams, TestCrate};
+use test_collector::{RawForkConfig, RawForkParams};
 
-use scarb_artifacts::StarknetContractArtifacts;
-
-use crate::fuzzer::RandomFuzzer;
-use crate::scarb::config::ForkTarget;
-
-pub use crate::collecting::{collect_test_compilation_targets, TestCompilationTarget};
-pub use crate::test_crate_summary::TestCrateSummary;
-
-use crate::collecting::{
-    compile_tests, CompiledTestCrate, CompiledTestCrateRaw, CompiledTestCrateRunnable,
-    TestCaseRunnable, ValidatedForkConfig,
-};
+use crate::collecting::{collect_test_compilation_targets, compile_tests, CompiledTestCrateRaw};
 use crate::test_filter::TestsFilter;
-use test_collector::{FuzzerConfig, LinkedLibrary, RawForkConfig, RawForkParams, TestCase};
 
+pub mod collecting;
 pub mod pretty_printing;
 pub mod scarb;
-pub mod test_case_summary;
 pub mod test_filter;
-
-mod collecting;
-mod fuzzer;
-mod running;
-mod test_crate_summary;
-mod test_execution_syscall_handler;
-
-pub const FUZZER_RUNS_DEFAULT: u32 = 256;
-pub const CACHE_DIR: &str = ".snfoundry_cache";
-
-static BUILTINS: Lazy<Vec<&str>> = Lazy::new(|| {
-    vec![
-        "Pedersen",
-        "RangeCheck",
-        "Bitwise",
-        "EcOp",
-        "Poseidon",
-        "SegmentArena",
-        "GasBuiltin",
-        "System",
-    ]
-});
-
-/// Configuration of the test runner
-#[derive(Debug, PartialEq)]
-pub struct RunnerConfig {
-    pub workspace_root: Utf8PathBuf,
-    pub exit_first: bool,
-    pub fork_targets: Vec<ForkTarget>,
-    pub fuzzer_runs: u32,
-    pub fuzzer_seed: u64,
-}
-
-impl RunnerConfig {
-    /// Creates a new `RunnerConfig` from given arguments
-    ///
-    /// # Arguments
-    ///
-    /// * `test_name_filter` - Used to filter test cases by names
-    /// * `exact_match` - Should test names match the `test_name_filter` exactly
-    /// * `exit_first` - Should runner exit after first failed test
-    #[must_use]
-    #[allow(clippy::too_many_arguments, clippy::fn_params_excessive_bools)]
-    pub fn new(
-        workspace_root: Utf8PathBuf,
-        exit_first: bool,
-        fork_targets: Vec<ForkTarget>,
-        fuzzer_runs: u32,
-        fuzzer_seed: u64,
-    ) -> Self {
-        Self {
-            workspace_root,
-            exit_first,
-            fork_targets,
-            fuzzer_runs,
-            fuzzer_seed,
-        }
-    }
-}
-
-pub struct CancellationTokens {
-    exit_first: CancellationToken,
-    error: CancellationToken,
-}
-
-impl CancellationTokens {
-    #[must_use]
-    pub fn new() -> Self {
-        let exit_first = CancellationToken::new();
-        let error = CancellationToken::new();
-        Self { exit_first, error }
-    }
-}
-
-impl Default for CancellationTokens {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Exit status of the runner
-#[derive(Debug, PartialEq, Clone)]
-pub enum RunnerStatus {
-    /// Runner exited without problems
-    Default,
-    /// Some test failed
-    TestFailed,
-    /// Runner did not run, e.g. when test cases got skipped
-    DidNotRun,
-}
-
-pub struct RunnerParams {
-    corelib_path: Utf8PathBuf,
-    contracts: HashMap<String, StarknetContractArtifacts>,
-    predeployed_contracts: Utf8PathBuf,
-    environment_variables: HashMap<String, String>,
-    linked_libraries: Vec<LinkedLibrary>,
-}
-
-impl RunnerParams {
-    #[must_use]
-    pub fn new(
-        corelib_path: Utf8PathBuf,
-        contracts: HashMap<String, StarknetContractArtifacts>,
-        predeployed_contracts: Utf8PathBuf,
-        environment_variables: HashMap<String, String>,
-        linked_libraries: Vec<LinkedLibrary>,
-    ) -> Self {
-        Self {
-            corelib_path,
-            contracts,
-            predeployed_contracts,
-            environment_variables,
-            linked_libraries,
-        }
-    }
-}
 
 #[derive(Debug, PartialEq, Clone, Copy)]
 pub enum CrateLocation {
@@ -171,8 +25,8 @@ pub enum CrateLocation {
     Tests,
 }
 
-fn parse_fork_params(raw_fork_params: &RawForkParams) -> Result<ValidatedForkConfig> {
-    Ok(ValidatedForkConfig {
+fn parse_fork_params(raw_fork_params: &RawForkParams) -> Result<ForkConfig> {
+    Ok(ForkConfig {
         url: raw_fork_params.url.parse()?,
         block_id: raw_fork_params.block_id,
     })
@@ -201,19 +55,19 @@ fn replace_id_with_params(
 fn to_runnable(
     compiled_test_crate: CompiledTestCrateRaw,
     runner_config: &RunnerConfig,
-) -> Result<CompiledTestCrateRunnable> {
+) -> Result<TestCrate> {
     let mut test_cases = vec![];
 
     for case in compiled_test_crate.test_cases {
         let fork_config = if let Some(fc) = case.fork_config {
             let raw_fork_params = replace_id_with_params(fc, runner_config)?;
-            let validated_fork_config = parse_fork_params(&raw_fork_params)?;
-            Some(validated_fork_config)
+            let fork_config = parse_fork_params(&raw_fork_params)?;
+            Some(fork_config)
         } else {
             None
         };
 
-        test_cases.push(TestCase {
+        test_cases.push(RunnerTestCase {
             name: case.name,
             available_gas: case.available_gas,
             ignored: case.ignored,
@@ -223,10 +77,9 @@ fn to_runnable(
         });
     }
 
-    Ok(CompiledTestCrate {
+    Ok(TestCrate {
         sierra_program: compiled_test_crate.sierra_program,
         test_cases,
-        tests_location: compiled_test_crate.tests_location,
     })
 }
 
@@ -260,10 +113,6 @@ pub async fn run(
         .into_iter()
         .map(|tc| tests_filter.filter_tests(tc))
         .collect_vec();
-    let test_crates = test_crates
-        .into_iter()
-        .map(|ctc| to_runnable(ctc, &runner_config))
-        .collect::<Result<Vec<_>>>()?;
 
     pretty_printing::print_collected_tests_count(
         test_crates.iter().map(|tests| tests.test_cases.len()).sum(),
@@ -273,18 +122,19 @@ pub async fn run(
     let mut summaries = vec![];
 
     for compiled_test_crate in test_crates {
-        let compiled_test_crate = Arc::new(compiled_test_crate);
-        let runner_config = runner_config.clone();
-        let runner_params = runner_params.clone();
-        let cancellation_tokens = cancellation_tokens.clone();
-
         pretty_printing::print_running_tests(
             compiled_test_crate.tests_location,
             compiled_test_crate.test_cases.len(),
         );
 
-        let summary = run_tests_from_crate(
-            compiled_test_crate,
+        let compiled_test_crate = to_runnable(compiled_test_crate, &runner_config)?;
+        let compiled_test_crate = Arc::new(compiled_test_crate);
+        let runner_config = runner_config.clone();
+        let runner_params = runner_params.clone();
+        let cancellation_tokens = cancellation_tokens.clone();
+
+        let summary = forge_runner::run_tests_from_crate(
+            compiled_test_crate.clone(),
             runner_config,
             runner_params,
             cancellation_tokens,
@@ -307,344 +157,25 @@ pub async fn run(
     Ok(summaries)
 }
 
-pub trait TestCaseFilter {
-    fn should_be_run(&self, test_case: &TestCase<ValidatedForkConfig>) -> bool;
-}
-
-async fn run_tests_from_crate(
-    tests: Arc<CompiledTestCrateRunnable>,
-    runner_config: Arc<RunnerConfig>,
-    runner_params: Arc<RunnerParams>,
-    cancellation_tokens: Arc<CancellationTokens>,
-    tests_filter: &impl TestCaseFilter,
-) -> Result<TestCrateSummary> {
-    let runner = Arc::new(
-        SierraCasmRunner::new(
-            tests.sierra_program.clone(),
-            Some(MetadataComputationConfig::default()),
-            OrderedHashMap::default(),
-        )
-        .context("Failed setting up runner.")?,
-    );
-
-    let mut tasks = FuturesUnordered::new();
-    let test_cases = &tests.test_cases;
-    // Initiate two channels to manage the `--exit-first` flag.
-    // Owing to `cheatnet` fork's utilization of its own Tokio runtime for RPC requests,
-    // test execution must occur within a `tokio::spawn_blocking`.
-    // As `spawn_blocking` can't be prematurely cancelled (refer: https://dtantsur.github.io/rust-openstack/tokio/task/fn.spawn_blocking.html),
-    // a channel is used to signal the task that test processing is no longer necessary.
-    let (send, mut rec) = channel(1);
-
-    // The second channel serves as a hold point to ensure all tasks complete
-    // their shutdown procedures before moving forward (more info: https://tokio.rs/tokio/topics/shutdown)
-    let (send_shut_down, mut rec_shut_down) = channel(1);
-
-    for case in test_cases.iter() {
-        let case_name = case.name.clone();
-
-        if !tests_filter.should_be_run(case) {
-            tasks.push(tokio::task::spawn(async {
-                Ok(TestCaseSummary::Ignored { name: case_name })
-            }));
-            continue;
-        };
-
-        let function = runner.find_function(&case_name)?;
-        let args = function_args(function, &BUILTINS);
-
-        let case = Arc::new(case.clone());
-        let args: Vec<ConcreteTypeId> = args.into_iter().cloned().collect();
-        let runner = runner.clone();
-
-        tasks.push(choose_test_strategy_and_run(
-            args,
-            case.clone(),
-            runner,
-            runner_config.clone(),
-            runner_params.clone(),
-            cancellation_tokens.clone(),
-            &send,
-            &send_shut_down,
-        ));
-    }
-
-    let mut results = vec![];
-
-    while let Some(task) = tasks.next().await {
-        let result = task??;
-
-        pretty_printing::print_test_result(&result);
-        results.push(result);
-    }
-
-    rec.close();
-
-    // Waiting for things to finish shutting down
-    drop(send_shut_down);
-    let _ = rec_shut_down.recv().await;
-
-    let contained_fuzzed_tests = results.iter().any(|summary| summary.runs().is_some());
-    Ok(TestCrateSummary {
-        test_case_summaries: results,
-        runner_exit_status: RunnerStatus::Default,
-        test_crate_type: tests.tests_location,
-        contained_fuzzed_tests,
-    })
-}
-
-#[allow(clippy::too_many_arguments)]
-fn choose_test_strategy_and_run(
-    args: Vec<ConcreteTypeId>,
-    case: Arc<TestCaseRunnable>,
-    runner: Arc<SierraCasmRunner>,
-    runner_config: Arc<RunnerConfig>,
-    runner_params: Arc<RunnerParams>,
-    cancellation_tokens: Arc<CancellationTokens>,
-    send: &Sender<()>,
-    send_shut_down: &Sender<()>,
-) -> JoinHandle<Result<TestCaseSummary>> {
-    if args.is_empty() {
-        run_single_test(
-            case,
-            runner,
-            runner_config,
-            runner_params,
-            cancellation_tokens,
-            send.clone(),
-            send_shut_down.clone(),
-        )
-    } else {
-        run_with_fuzzing(
-            args,
-            case,
-            runner,
-            runner_config,
-            runner_params,
-            cancellation_tokens,
-            send_shut_down.clone(),
-        )
-    }
-}
-
-fn run_single_test(
-    case: Arc<TestCaseRunnable>,
-    runner: Arc<SierraCasmRunner>,
-    runner_config: Arc<RunnerConfig>,
-    runner_params: Arc<RunnerParams>,
-    cancellation_tokens: Arc<CancellationTokens>,
-    send: Sender<()>,
-    send_shut_down: Sender<()>,
-) -> JoinHandle<Result<TestCaseSummary>> {
-    let exit_first = runner_config.exit_first;
-    tokio::task::spawn(async move {
-        tokio::select! {
-            () = cancellation_tokens.exit_first.cancelled() => {
-                // Stop executing all tests because flag --exit-first'
-                // has been set and one test FAIL
-                Ok(TestCaseSummary::Skipped {})
-            },
-            () = cancellation_tokens.error.cancelled() => {
-                // Stop executing all tests because
-                // one of a test returns Err
-                Ok(TestCaseSummary::Skipped {})
-            },
-
-            result = blocking_run_from_test(vec![], case.clone(),runner,  runner_config.clone(), runner_params.clone(), send.clone(), send_shut_down.clone() ) => {
-                match result? {
-                    Ok(result) => {
-                        if exit_first {
-                            if let TestCaseSummary::Failed { .. } = &result {
-                                cancellation_tokens.exit_first.cancel();
-                            }
-                        }
-                        Ok(result)
-                    }
-                    Err(e) => {
-                        cancellation_tokens.error.cancel();
-                        Err(e)
-                    }
-                }
-            }
-        }
-    })
-}
-
-fn run_with_fuzzing(
-    args: Vec<ConcreteTypeId>,
-    case: Arc<TestCaseRunnable>,
-    runner: Arc<SierraCasmRunner>,
-    runner_config: Arc<RunnerConfig>,
-    runner_params: Arc<RunnerParams>,
-    cancellation_tokens: Arc<CancellationTokens>,
-    send_shut_down: Sender<()>,
-) -> JoinHandle<Result<TestCaseSummary>> {
-    tokio::task::spawn(async move {
-        let cancellation_fuzzing_token = CancellationToken::new();
-        let (send, mut rec) = channel(1);
-        let args = args
-            .iter()
-            .map(|arg| {
-                arg.debug_name
-                    .as_ref()
-                    .ok_or_else(|| anyhow!("Type {arg:?} does not have a debug name"))
-                    .map(SmolStr::as_str)
-            })
-            .collect::<Result<Vec<_>>>()?;
-
-        let (fuzzer_runs, fuzzer_seed) = match case.fuzzer_config {
-            Some(FuzzerConfig {
-                fuzzer_runs,
-                fuzzer_seed,
-            }) => (fuzzer_runs, fuzzer_seed),
-            _ => (runner_config.fuzzer_runs, runner_config.fuzzer_seed),
-        };
-        let mut fuzzer = RandomFuzzer::create(fuzzer_seed, fuzzer_runs, &args)?;
-
-        let mut tasks = FuturesUnordered::new();
-
-        for _ in 1..=fuzzer_runs {
-            let args = fuzzer.next_args();
-
-            tasks.push(run_fuzzing_subtest(
-                args,
-                case.clone(),
-                runner.clone(),
-                runner_config.clone(),
-                runner_params.clone(),
-                cancellation_tokens.clone(),
-                cancellation_fuzzing_token.clone(),
-                send.clone(),
-                send_shut_down.clone(),
-            ));
-        }
-
-        let mut results = vec![];
-        let mut final_result = None;
-
-        while let Some(task) = tasks.next().await {
-            let result = task??;
-
-            results.push(result.clone());
-            final_result = Some(result.clone());
-
-            if let TestCaseSummary::Failed { .. } = result {
-                cancellation_fuzzing_token.cancel();
-                break;
-            }
-        }
-
-        rec.close();
-
-        let runs = u32::try_from(
-            results
-                .iter()
-                .filter(|item| {
-                    matches!(
-                        item,
-                        TestCaseSummary::Passed { .. } | TestCaseSummary::Failed { .. }
-                    )
-                })
-                .count(),
-        )?;
-
-        match final_result {
-            Some(result) => Ok(result.with_runs(runs)),
-            None => panic!("Test should always run at least once"),
-        }
-    })
-}
-
-#[allow(clippy::too_many_arguments)]
-fn run_fuzzing_subtest(
-    args: Vec<Felt252>,
-    case: Arc<TestCaseRunnable>,
-    runner: Arc<SierraCasmRunner>,
-    runner_config: Arc<RunnerConfig>,
-    runner_params: Arc<RunnerParams>,
-    cancellation_tokens: Arc<CancellationTokens>,
-    cancellation_fuzzing_token: CancellationToken,
-    send: Sender<()>,
-    send_shut_down: Sender<()>,
-) -> JoinHandle<Result<TestCaseSummary>> {
-    task::spawn(async move {
-        tokio::select! {
-            () = cancellation_tokens.error.cancelled() => {
-                // Stop executing all tests because
-                // one of a test returns Err
-                Ok(TestCaseSummary::Skipped {  })
-            },
-            () = cancellation_tokens.exit_first.cancelled() => {
-                // Stop executing all tests because flag --exit-first'
-                // has been set and one test FAIL
-                Ok(TestCaseSummary::Skipped {  })
-            },
-            () = cancellation_fuzzing_token.cancelled() => {
-                // Stop executing all single fuzzing tests
-                // because one of fuzzing test has been FAIL
-                Ok(TestCaseSummary::Skipped {  })
-
-            },
-           result = blocking_run_from_test(
-                args.clone(),
-                case,
-                runner,
-                runner_config.clone(),
-                runner_params.clone(),
-                send.clone(),
-                send_shut_down.clone()
-            ) => {
-                match result? {
-                    Ok(result) => {
-                        if let TestCaseSummary::Failed { .. } = &result {
-                            if runner_config.exit_first {
-                                cancellation_tokens.exit_first.cancel();
-                            }
-                        }
-                        Ok(result)
-                    }
-                    Err(e) => {
-                        cancellation_tokens.error.cancel();
-                        Err(e)
-                    }
-                }
-            },
-        }
-    })
-}
-
-fn function_args<'a>(function: &'a Function, builtins: &[&str]) -> Vec<&'a ConcreteTypeId> {
-    let builtins: Vec<_> = builtins
-        .iter()
-        .map(|builtin| Some(SmolStr::new(builtin)))
-        .collect();
-
-    function
-        .signature
-        .param_types
-        .iter()
-        .filter(|pt| !builtins.contains(&pt.debug_name))
-        .collect()
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use cairo_lang_sierra::program::Program;
+    use forge_runner::ForkTarget;
     use starknet::core::types::BlockId;
     use starknet::core::types::BlockTag::Latest;
-    use test_collector::ExpectedTestResult;
+    use test_collector::{ExpectedTestResult, TestCase};
 
     #[test]
     fn to_runnable_unparsable_url() {
-        let mocked_tests = CompiledTestCrate {
+        let mocked_tests = CompiledTestCrateRaw {
             sierra_program: Program {
                 type_declarations: vec![],
                 libfunc_declarations: vec![],
                 statements: vec![],
                 funcs: vec![],
             },
-            test_cases: vec![TestCase {
+            test_cases: vec![TestCase::<RawForkConfig> {
                 name: "crate1::do_thing".to_string(),
                 available_gas: None,
                 ignored: false,
@@ -664,14 +195,14 @@ mod tests {
 
     #[test]
     fn to_runnable_non_existent_id() {
-        let mocked_tests = CompiledTestCrate {
+        let mocked_tests = CompiledTestCrateRaw {
             sierra_program: Program {
                 type_declarations: vec![],
                 libfunc_declarations: vec![],
                 statements: vec![],
                 funcs: vec![],
             },
-            test_cases: vec![TestCase {
+            test_cases: vec![TestCase::<RawForkConfig> {
                 name: "crate1::do_thing".to_string(),
                 available_gas: None,
                 ignored: false,
