+++ crates/forge/src/pretty_printing.rs
@@ -1,7 +1,8 @@
-use crate::test_case_summary::TestCaseSummary;
-use crate::{CrateLocation, TestCrateSummary};
+use crate::CrateLocation;
 use anyhow::Error;
 use console::style;
+use forge_runner::test_case_summary::TestCaseSummary;
+use forge_runner::test_crate_summary::TestCrateSummary;
 
 pub fn print_error_message(error: &Error) {
     let error_tag = style("ERROR").red();
@@ -39,57 +40,6 @@ pub(crate) fn print_test_summary(summaries: &[TestCrateSummary]) {
 
 pub(crate) fn print_test_seed(seed: u64) {
     println!("{}: {seed}", style("Fuzzer seed").bold());
-}
-
-pub(crate) fn print_test_result(test_result: &TestCaseSummary) {
-    if let TestCaseSummary::Skipped { .. } = test_result {
-        return;
-    }
-
-    let result_header = match test_result {
-        TestCaseSummary::Passed { .. } => format!("[{}]", style("PASS").green()),
-        TestCaseSummary::Failed { .. } => format!("[{}]", style("FAIL").red()),
-        TestCaseSummary::Ignored { .. } => format!("[{}]", style("IGNORE").yellow()),
-        TestCaseSummary::Skipped { .. } => {
-            unreachable!()
-        }
-    };
-
-    let result_name = match test_result {
-        TestCaseSummary::Ignored { name }
-        | TestCaseSummary::Failed { name, .. }
-        | TestCaseSummary::Passed { name, .. } => name,
-        TestCaseSummary::Skipped {} => {
-            unreachable!()
-        }
-    };
-
-    let result_message = match test_result {
-        TestCaseSummary::Passed { msg: Some(msg), .. } => format!("\n\nSuccess data:{msg}"),
-        TestCaseSummary::Failed { msg: Some(msg), .. } => format!("\n\nFailure data:{msg}"),
-        _ => String::new(),
-    };
-
-    let fuzzer_report = match test_result.runs() {
-        None => String::new(),
-        Some(runs) => {
-            if matches!(test_result, TestCaseSummary::Failed { .. }) {
-                let arguments = test_result.arguments();
-                format!(" (fuzzer runs = {runs}, arguments = {arguments:?})")
-            } else {
-                format!(" (fuzzer runs = {runs})")
-            }
-        }
-    };
-
-    let block_number_message = match test_result.latest_block_number() {
-        None => String::new(),
-        Some(latest_block_number) => {
-            format!("\nNumber of the block used for fork testing = {latest_block_number}")
-        }
-    };
-
-    println!("{result_header} {result_name}{fuzzer_report}{block_number_message}{result_message}");
 }
 
 pub fn print_failures(all_failed_tests: &[TestCaseSummary]) {
