+++ crates/forge/src/test_filter.rs
@@ -1,6 +1,5 @@
-use crate::collecting::{CompiledTestCrate, CompiledTestCrateRaw, ValidatedForkConfig};
-use crate::TestCaseFilter;
-use test_collector::TestCase;
+use crate::collecting::{CompiledTestCrate, CompiledTestCrateRaw};
+use forge_runner::{TestCase, TestCaseFilter};
 
 #[derive(Debug, PartialEq)]
 // Specifies what tests should be included
@@ -85,7 +84,7 @@ impl TestsFilter {
 }
 
 impl TestCaseFilter for TestsFilter {
-    fn should_be_run(&self, test_case: &TestCase<ValidatedForkConfig>) -> bool {
+    fn should_be_run(&self, test_case: &TestCase) -> bool {
         match self.ignored_filter {
             IgnoredFilter::All => true,
             IgnoredFilter::Ignored => test_case.ignored,
