+++ crates/forge/test_utils/Cargo.toml
@@ -26,3 +26,4 @@ tokio.workspace = true
 forge = { path = "../../forge" }
 test-collector = { path = "../../test-collector" }
 scarb-artifacts = { path = "../../scarb-artifacts" }
+forge_runner = { path = "../../forge_runner" }
