+++ crates/forge/tests/e2e/forking.rs
@@ -2,7 +2,7 @@ use crate::assert_stdout_contains;
 use crate::e2e::common::runner::{
     runner, setup_package, setup_package_with_file_patterns, test_runner, BASE_FILE_PATTERNS,
 };
-use forge::CACHE_DIR;
+use forge_runner::CACHE_DIR;
 use indoc::indoc;
 
 #[test]
