+++ crates/forge_runner/Cargo.toml
@@ -1,0 +1,43 @@
+[package]
+name = "forge_runner"
+version.workspace = true
+edition.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow.workspace = true
+cairo-lang-runner.workspace = true
+cairo-lang-casm.workspace = true
+cairo-lang-sierra-to-casm.workspace = true
+cairo-lang-sierra.workspace = true
+cairo-lang-utils.workspace = true
+cairo-lang-starknet.workspace = true
+cairo-lang-compiler.workspace = true
+cairo-lang-filesystem.workspace = true
+cairo-felt.workspace = true
+starknet_api.workspace = true
+starknet.workspace = true
+futures.workspace = true
+once_cell.workspace = true
+tokio.workspace = true
+tokio-util.workspace = true
+smol_str = "0.2.0"
+num-integer = "0.1.45"
+num-bigint.workspace = true
+num-traits.workspace = true
+rand.workspace = true
+url.workspace = true
+blockifier.workspace = true
+cairo-vm = "0.8.1"
+itertools.workspace = true
+camino.workspace = true
+thiserror.workspace = true
+serde_json.workspace = true
+console.workspace = true
+serde.workspace = true
+flatten-serde-json = "0.1.0"
+cheatnet = { path = "../cheatnet" }
+conversions = { path = "../conversions" }
+scarb-artifacts = { path = "../scarb-artifacts" }
+test-collector = { path = "../test-collector" }
