+++ crates/forge_runner/src/fuzzer.rs
@@ -1,7 +1,8 @@
-use crate::fuzzer::arguments::CairoType;
 use anyhow::Result;
 use rand::rngs::StdRng;
 use rand::Rng;
+
+use crate::fuzzer::arguments::CairoType;
 
 mod arguments;
 mod random;
