+++ crates/forge_runner/src/test_case_summary.rs
@@ -1,5 +1,5 @@
-use crate::collecting::TestCaseRunnable;
 use crate::running::ForkInfo;
+use crate::TestCase;
 use cairo_felt::Felt252;
 use cairo_lang_runner::short_string::as_cairo_short_string;
 use cairo_lang_runner::{RunResult, RunResultValue};
@@ -51,14 +51,16 @@ pub enum TestCaseSummary {
 }
 
 impl TestCaseSummary {
-    pub(crate) fn arguments(&self) -> Vec<Felt252> {
+    #[must_use]
+    pub fn arguments(&self) -> Vec<Felt252> {
         match self {
             TestCaseSummary::Failed { arguments, .. }
             | TestCaseSummary::Passed { arguments, .. } => arguments.clone(),
             TestCaseSummary::Ignored { .. } | TestCaseSummary::Skipped { .. } => vec![],
         }
     }
-    pub(crate) fn runs(&self) -> Option<u32> {
+    #[must_use]
+    pub fn runs(&self) -> Option<u32> {
         match self {
             TestCaseSummary::Failed {
                 fuzzing_statistic, ..
@@ -86,7 +88,8 @@ impl TestCaseSummary {
         }
     }
 
-    pub(crate) fn with_runs(self, runs: u32) -> Self {
+    #[must_use]
+    pub fn with_runs(self, runs: u32) -> Self {
         match self {
             TestCaseSummary::Passed {
                 name,
@@ -124,7 +127,7 @@ impl TestCaseSummary {
     #[must_use]
     pub(crate) fn from_run_result_and_info(
         run_result: RunResult,
-        test_case: &TestCaseRunnable,
+        test_case: &TestCase,
         arguments: Vec<Felt252>,
         fork_info: &ForkInfo,
     ) -> Self {
@@ -202,7 +205,7 @@ fn build_readable_text(data: &Vec<Felt252>) -> Option<String> {
 /// Returns a string with the data that was produced by the test case.
 /// If the test was expected to fail with specific data e.g. `#[should_panic(expected: ('data',))]`
 /// and failed to do so, it returns a string comparing the panic data and the expected data.
-pub(crate) fn extract_result_data(
+pub fn extract_result_data(
     run_result: &RunResult,
     expectation: &ExpectedTestResult,
 ) -> Option<String> {
