+++ crates/forge_runner/src/test_crate_summary.rs
@@ -1,5 +1,5 @@
 use crate::test_case_summary::TestCaseSummary;
-use crate::{CrateLocation, RunnerStatus};
+use crate::RunnerStatus;
 
 /// Summary of the test run in the file
 #[derive(Debug, PartialEq)]
@@ -8,28 +8,31 @@ pub struct TestCrateSummary {
     pub test_case_summaries: Vec<TestCaseSummary>,
     /// Status of the runner after executing tests in the file
     pub runner_exit_status: RunnerStatus,
-    /// Type of the test crate
-    pub test_crate_type: CrateLocation,
+    // /// Type of the test crate
+    // pub test_crate_type: CrateLocation,
     /// If test crate contained fuzzed tests
     pub contained_fuzzed_tests: bool,
 }
 
 impl TestCrateSummary {
-    pub(crate) fn count_passed(&self) -> usize {
+    #[must_use]
+    pub fn count_passed(&self) -> usize {
         self.test_case_summaries
             .iter()
             .filter(|tu| matches!(tu, TestCaseSummary::Passed { .. }))
             .count()
     }
 
-    pub(crate) fn count_failed(&self) -> usize {
+    #[must_use]
+    pub fn count_failed(&self) -> usize {
         self.test_case_summaries
             .iter()
             .filter(|tu| matches!(tu, TestCaseSummary::Failed { .. }))
             .count()
     }
 
-    pub(crate) fn count_skipped(&self) -> usize {
+    #[must_use]
+    pub fn count_skipped(&self) -> usize {
         self.test_case_summaries
             .iter()
             .filter(|tu| matches!(tu, TestCaseSummary::Skipped { .. }))
